refactor(test): share bootstrap dependency mocks
